Tdev API DEV Rap/Tdev-API/requirements.txt
fastapi[all]==0.92.0
gunicorn==20.1.0
pydantic==1.7.4
pymongo==4.3.3
azure-storage-blob==12.9.0
azure-storage-queue==12.5.0
tqdm==4.63.0
pandas==1.3.5
requests==2.27.1
numpy==1.24.1
pycoingecko==3.1.0
tk==0.1.0
openpyxl==3.0.9
python-jose==3.3.0
opencensus-ext-azure==1.1.8